+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYVW8Z849WJKEZ6QXCT0KNAE"
+        "01KYW65MMPXCFB94VCSSBE81H0"
     )
     r"""The unique identifier of the data source"""