+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -40,9 +40,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-31T10:42:02.740Z"
+    created: Optional[str] = "2026-07-31T13:34:59.403Z"
 
-    updated: Optional[str] = "2026-07-31T10:42:02.740Z"
+    updated: Optional[str] = "2026-07-31T13:34:59.403Z"
 
     updated_by_id: OptionalNullable[str] = UNSET