+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYVW8Z81X1AWEJ35K2JAMBK7"
+        "01KYW65MMKHQ6V9JT55QBG3AES"
     )
     r"""The unique identifier of the data source"""