+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYVW8Z835ZNTFR085N6VFR1V"
+        "01KYW65MMNZ786CQ4PK7W9Z9FQ"
     )
     r"""The unique identifier of the data source"""