+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-31T10:42:00.647Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-31T13:34:57.400Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")