+++ packages/orq-rc/RELEASES.md
@@ -9909,3 +9909,13 @@ Based on:
 - [python v4.13.0-rc.24] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.24 - packages/orq-rc
+
+## 2026-07-31 13:38:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.25 - packages/orq-rc